+++ tests/storages_basic.rs
@@ -35,23 +35,16 @@ impl TryFrom<String> for SessionData {
         Ok(Self { user_id: value })
     }
 }
+impl From<SessionData> for String {
+    fn from(value: SessionData) -> Self {
+        value.user_id
+    }
+}
 impl fred::types::FromValue for SessionData {
     fn from_value(value: fred::prelude::Value) -> Result<Self, fred::prelude::Error> {
         Ok(Self {
             user_id: value.convert()?,
         })
-    }
-}
-impl std::fmt::Display for SessionData {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.user_id)
-    }
-}
-impl TryFrom<fred::types::Value> for SessionData {
-    type Error = SessionError;
-    fn try_from(value: fred::types::Value) -> Result<Self, Self::Error> {
-        let user_id = value.as_string().ok_or(SessionError::NotFound)?;
-        Ok(Self { user_id })
     }
 }
 impl From<SessionData> for fred::types::Value {
